+++ solana/bridge_token_factory/programs/bridge_token_factory/src/state/config.rs
@@ -19,7 +19,7 @@ pub struct ConfigBumps {
 #[derive(InitSpace)]
 pub struct Config {
     pub admin: Pubkey,
-    pub max_used_nonce: u128,
+    pub max_used_nonce: u64,
     pub derived_near_bridge_address: [u8; 64],
     pub bumps: ConfigBumps,
 }
